v2.1.0 - Maintenance release and product update - see Release_Notes.html

+++ Projects/STM32U585AI-STWIN.box/Applications/DATALOGMC_AI/Core/Src/OTA.c
@@ -75,7 +75,7 @@ static FwId_BoardName_t CurrentFwInfo, OtherBankFwInfo;
 static void DeleteOtherFlashBankFwId(void);
 
 /* Exported functions --------------------------------------------------------*/
-extern int32_t CurrentActiveBank;
+extern uint8_t CurrentActiveBank;
 /**
   * @brief Function for Updating the Firmware
   * @param uint32_t *SizeOfUpdate Remaining size of the firmware image [bytes]
